// frontend/src/app/api/model/companyViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CompanyViewDto { 
    id?: number;
    name?: string;
    registrationNumber?: string;
}