// frontend/src/app/api/model/officeDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OperatingHourDto } from './operatingHourDto';
import { AddressDetailsDto } from './addressDetailsDto';


export interface OfficeDto { 
    companyId: number;
    address: AddressDetailsDto;
    operatingHours: Set<OperatingHourDto>;
}