// frontend/src/app/api/model/officeViewDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OperatingHourViewDto } from './operatingHourViewDto';


export interface OfficeViewDto { 
    id?: number;
    companyId?: number;
    cityName?: string;
    cityPostcode?: string;
    fullAddress?: string;
    operatingHours?: Set<OperatingHourViewDto>;
}