// frontend/src/app/api/model/operatingHourDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OperatingHourDto { 
    dayOfWeek: OperatingHourDto.DayOfWeekEnum;
    openTime?: string;
    closeTime?: string;
    isClosed?: boolean;
}
export namespace OperatingHourDto {
    export const DayOfWeekEnum = {
        Monday: 'MONDAY',
        Tuesday: 'TUESDAY',
        Wednesday: 'WEDNESDAY',
        Thursday: 'THURSDAY',
        Friday: 'FRIDAY',
        Saturday: 'SATURDAY',
        Sunday: 'SUNDAY'
    } as const;
    export type DayOfWeekEnum = typeof DayOfWeekEnum[keyof typeof DayOfWeekEnum];
}